Update _index.md
Added short bio and contact information for myself. If I understand it correctly this will not appear at the website, but can be used for a practice. I am not sure though how should this file look like if we wanted to include both Rick's and my info to it, and how would it display on the actual web page.